tests: Move `GET /summary` tests to `krate::summary` module
